Autenticação pronta

+++ gerenciador_tarefas/app/views/tarefa_views.py
@@ -1,15 +1,19 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from ..forms import TarefaForm
 from ..entidades.tarefa import Tarefa
 from ..services import tarefa_service
 
 # Create your views here.
 
+@login_required()
 def listar_tarefas(request):
-    tarefas = tarefa_service.listar_tarefas()
+    tarefas = tarefa_service.listar_tarefas(request.user)
 
     return render(request, 'tarefas/listar_tarefas.html', {'tarefas': tarefas})
 
+@login_required()
 def cadastrar_tarefa(request):
     if request.method == 'POST':
         form_tarefa = TarefaForm(request.POST)
@@ -19,10 +23,16 @@ def cadastrar_tarefa(request):
             descricao = form_tarefa.cleaned_data['descricao']
             data_expiracao = form_tarefa.cleaned_data['data_expiracao']
             prioridade = form_tarefa.cleaned_data['prioridade']
+            usuario = request.user
 
-            print(titulo)
 
-            tarefa_nova = Tarefa(titulo=titulo, descricao=descricao, data_expiracao=data_expiracao, prioridade=prioridade)
+            tarefa_nova = Tarefa(
+                titulo = titulo,
+                descricao = descricao,
+                data_expiracao = data_expiracao,
+                prioridade = prioridade,
+                usuario = usuario
+            )
             print(tarefa_nova.titulo)
             tarefa_service.cadastrar_tarefa(tarefa_nova)
 
@@ -32,8 +42,13 @@ def cadastrar_tarefa(request):
 
     return render(request, 'tarefas/form_tarefa.html', {'form_tarefa': form_tarefa})
 
+@login_required()
 def editar_tarefa(request, id):
     tarefa_bd = tarefa_service.listar_tarefa_id(id)
+
+    if tarefa_bd.usuario != request.user:
+        return httpResponse('Não permitido')
+
     form_tarefa = TarefaForm(request.POST or None, instance = tarefa_bd)
 
     if form_tarefa.is_valid():
@@ -41,10 +56,9 @@ def editar_tarefa(request, id):
         descricao = form_tarefa.cleaned_data['descricao']
         data_expiracao = form_tarefa.cleaned_data['data_expiracao']
         prioridade = form_tarefa.cleaned_data['prioridade']
+        usuario = request.user
 
-        print(titulo)
-
-        tarefa_nova = Tarefa(titulo=titulo, descricao=descricao, data_expiracao=data_expiracao, prioridade=prioridade)
+        tarefa_nova = Tarefa(titulo=titulo, descricao=descricao, data_expiracao=data_expiracao, prioridade=prioridade, usuario = usuario)
         print(tarefa_nova.titulo)
 
         tarefa_service.editar_tarefa(tarefa_bd, tarefa_nova)
@@ -53,8 +67,13 @@ def editar_tarefa(request, id):
 
     return render(request, 'tarefas/form_tarefa.html', {'form_tarefa': form_tarefa})
 
+@login_required()
 def remover_tarefa(request, id):
     tarefa_bd = tarefa_service.listar_tarefa_id(id)
+
+    if tarefa_bd.usuario != request.user:
+        return httpResponse('Não permitido')
+
     if request.method == 'POST':
         tarefa_service.remover_tarefa(tarefa_bd)
